application: don't try to start miners for account types we don't have
Right now we just start the miner processes, which will just return in
case no matching account is found.
Instead, only run a miner in case we know there's an account of a
certain type configured, and connect to signals for when accounts are
added or removed to check again.