preview: Simplify the PreviewNavButtons show/hide code
We could just call _updateVisibility instead of fading it in and out
ourself.

In fact, it would be more correct to do so because otherwise show will
always fade in both widgets -- which is wrong. The reason it works
correctly now is that show is called before setModel. So the call to
_fadeInButton from show aborts because this._model is null. This is
due to the particular order in which _onLoadFinished is called in
PreviewView and Embed.

https://bugzilla.gnome.org/show_bug.cgi?id=720985