preview, places: Disable bookmarking if metadata is unsupported
... without hiding all the navigation controls.

This is often due to the absence of GVfs. The absence of GVfs is
considered to be an example of a broken system, but to make the lives
of those who are building from Git into a separate prefix we try to
accommodate this without getting into elaborate fallback code paths.

https://bugzilla.gnome.org/show_bug.cgi?id=733061