Drop Gd.HeaderMenuButton
A Gd.HeaderMenuButton is a Gtk.MenuButton that either has a 16px icon,
tooltip text and the image-button style class; or a label and the
text-button style class.

Gtk.Button does not offer a use-markup property so we need to set it
directly on the child Gtk.Label. We also need to guard against NULL
strings because Gtk.Button.set_label is not marked as 'allow-none'.

https://bugzilla.gnome.org/show_bug.cgi?id=733683