Don't clear DocumentManager on every TrackerController refresh
We were clearing DocumentManager on every refresh of the
TrackerController. This will be problematic when we implement the new
split view designs because DocumentManager will have items spanning
across multiple views. Refreshing the TrackerController for one view
should not clear DocumentManager because it might have items for some
other view.

Instead, we can add some filtering to ViewModel so that it only picks
those items that are relevant to the view for which it is a model.

Since we are not clearing DocumentManager on every refresh, we will
find some items that are already part of the manager. Avoid emitting
an extra "object-removed" / "object-added" pair in these cases.

https://bugzilla.gnome.org/show_bug.cgi?id=686461