Don't show a glimpse of the older item when going back to the preview
Now that we switch to the preview as soon as an item starts loading, we
should take care to clear any older item that might have been previewed
earlier.

Since EvView does not allow unsetting the model (by passing NULL), we
destroy the old widget and create a new one. We do the same for the
NavControls because of GdNavBar.

https://bugzilla.gnome.org/show_bug.cgi?id=720516