windowMode: Clean up the fullscreen code
Ensure that the fullscreen logic runs after this._mode has been updated
so that there are no state changes between can-fullscreen-changed and
window-mode-changed.

This changes the current requirement that setFullscreen has to be used
before the mode is updated, which I found to be confusing, and we
don't need a separate this._canFullscreen variable any more.

https://bugzilla.gnome.org/show_bug.cgi?id=686461