search: Split OffsetController so that each view can have its own
When we have separate views for documents, collections and search, each
of them would require separate queries to count the number of items
they have, and the offset can be different in each case. So, we need to
allow each view to have its own OffsetController implementation.

Currently the only implementation is OffsetOverviewController, which is
the same as the unified OffsetController that we had.

https://bugzilla.gnome.org/show_bug.cgi?id=686461