trackerController: Split it so that each view can have its own
When we have multiple views, each of them would be showing different
kinds of items, and they will be reacting differently to changes in
search criteria. eg., the documents and collections views will always
show only documents and collections, while the search view will show a
mix of items depending on the current search criteria.

Therefore, we need to allow each view to have its own implementation.

Currently the only implementation is TrackerOverviewController, which
is the same as the unified TrackerController that we had.

https://bugzilla.gnome.org/show_bug.cgi?id=686461