Rework the application accommodate multiple views
The objective is to have separate views for documents, collections and
search results. The first two views will always show only documents and
collections arranged in reverse chronological order. Whenever search
constraints are applied the results are supposed to show up in the
search view.

Each view will have its separate TrackerController and OffsetController
implementations to decide when to refresh its contents and what the
current offsets are. The OffsetController is no longer a part of the
search context because the different views need to use their own
implementations. The search provider does not have a way to change the
offsets, so it can just use the initial value of [0, 50).

New flags corresponding to each of the views have been added to
QueryFlags. These are passed around to the classes that play a part in
constructing the queries.

The meaning of the NONE and UNFILTERED flags have also changed.
Earlier, UNFILTERED was used to completely turn off search categories,
search matches, search types and collections, while NONE had them
enabled. Now, NONE will disable the search type WHERE clauses but keep
everything else enabled so that it can be used to create queries for
browsing the contents of a collection.

Since the getFilter method of the various BaseManager classes are now
affected by their roles in the individual views, it no longer makes
sense to have a generic implementation in the parent. Instead,
getAllFilter has been made public so that the children can use it to
construct their query snippets.

ViewModel has grown a filtering mechanism to allow each view to only
have what is needed.

https://bugzilla.gnome.org/show_bug.cgi?id=686461