embed, view: Move the no-results page into the ViewContainer
With multiple views, it is simpler to give each view its own no-results
page, so that they can individually react to the presence or absence
of items.

Instead of listening to TrackerChangeMonitor, we use the
OffsetController and we reset the item count as new items are added to
ViewModel. This is useful because each view will have its own separate
OffsetController and ViewModel to filter the global item set, while
TrackerChangeMonitor has no such filtering and gets bombarded with
every change in the tracker database.

https://bugzilla.gnome.org/show_bug.cgi?id=686461