preview: fix showing of fullscreen toolbar
Since a recent change in libevinceview, the selection-changed signal is
now emitted on the EvinceView also when the selection hasn't actually
changed. This caused our logic to interpret every ckick in fullscreen as
"selection has changed", and we would never activate the fullscreen
toolbar, which is supposed to show on click when nothing else is
selected.
Cache the selection state in the view to work this around.

https://bugzilla.gnome.org/show_bug.cgi?id=725372