search: use g_str_tokenize_and_fold()
Use the GLib-provided g_str_tokenize_and_fold() function to tokenize our
string, instead of a hand-rolled function.
This changes the behavior slightly, in that the string we will use in
the filters will not be unaccented. We then duplicate filters so that
Tracker matches both on the accented and the unaccented string.

This means that e.g. "podminky" will return results that include both
"podminky" and "podmínky", but "podmínky" will only return results for
"podmínky". We assume that if the user typed an accent, it is in fact
intentional.

https://bugzilla.gnome.org/show_bug.cgi?id=722246