application: Spawn the miners only when a window is created
Creating the miner proxies during startup would spawn the processes
even though they are only used when an application window is shown.
This is bad because the startup is shared between the application and
the search provider, and we should not needlessly spawn so many
processes for each gnome-shell search.

https://bugzilla.gnome.org/show_bug.cgi?id=744625