application: Create the miner proxies asynchronously
We deliberately don't chain the creation of the proxies and assign the
variables (ie. this.gdataMiner, etc.) before leaving the main loop.
Otherwise, if the window is destroyed while the proxy creation was
still in flight, then some of the variables would get assigned after
they were null-ed, which would be a race.

Similarly, be careful about calling _refreshMinerNow with a null miner.

https://bugzilla.gnome.org/show_bug.cgi?id=744625