preview: Toggle the controls when the user taps on a touchscreen
Tapping on a touchscreen will toggle the visibility of the overlaid
controls without using an auto-hide timeout. The timeout is only used
when interacting with the mouse. Moving the mouse will continue to
reveal the controls and auto-hide them based on a timeout, as before.

To make the mouse and touchscreen interactions work together, the
queueing and unqueing of the timeout has been split out and used
explicitly where they are needed. An additional variable has been
introduced to control the visibility of the controls based on internal
conditions, as opposed to requests from external client code.

https://bugzilla.gnome.org/show_bug.cgi?id=741564