preview: Simplify the PreviewNavControls setModel code
There is no need to mix setting the model and updating the visibility
of the controls. Once a document has finished loading we set the model
and call the show method to reveal the controls. When we leave the
preview, we call hide or destroy the controls.

This is made possible by consolidating the load-finished code in the
previous patch.

https://bugzilla.gnome.org/show_bug.cgi?id=745507