preview: Simplify the load-finished logic
Instead of having PreviewView react to load-finished in two different
steps - once in its own handler, and then from Embed's handler, let's
consolidate everything in one place.

It is hard to know the order in which these handlers are getting
called. If it changed, the logic for showing or hiding the places and
bookmarks controls will break.

https://bugzilla.gnome.org/show_bug.cgi?id=745507