embed: Unbreak re-activation before the inactivity timeout
When we build the UI for the second time, we need to populate the
ViewModels with the items already present in the DocumentManager. We
lost this as a fall out from 4c7ffba09ac512c2814d4518cc2a63b61f942334

We also need to remove the rowRefs on each item when clearing the
application state. The rowRefs are used as markers to indicate if they
are already part of a ViewModel. If we don't clear them, then the
ViewModel's filtering mechanism will prevent the items from being
re-added when the application is re-activated.

https://bugzilla.gnome.org/show_bug.cgi?id=744643