Hide & restore search when moving to preview or collection view & back
We should hide the search bar when viewing a search result (ie. the
contents of a collection or previewing an item), and restore it when we
come back.

It is not clear whether the user is supposed to be able to search while
viewing a collection, so we ignore that case.

https://bugzilla.gnome.org/show_bug.cgi?id=746367