places: Hide "Contents" tab when no TOC is available
In the preview of a document, clicking the "Bookmarks" button
on the bottom left of the page shows a window with two tabs titled
'Contents' and 'Bookmarks' even if there is no table of contents
for the document. Now, the opened window doesn't show the Contents
tab if there is no TOC within the document.

https://bugzilla.gnome.org/show_bug.cgi?id=741780