mainToolbar: Disconnect from DocumentManager when destroyed
We destroy and re-create the toolbar on mode changes. Therefore, we
can not rely on disconnectAll to break all the connections when the
main window is finally gone. We need to break the connections
immediately when the toolbar is destroyed.

Failure to remove the signal handlers will lead to a spurious reference
being kept on the widget, which will prevent it from being garbage
collected.

https://bugzilla.gnome.org/show_bug.cgi?id=747506