preview: Fix possible warnings exiting fullscreen
After opening a document, turning on the fullscreen mode, and starting a
search (type a few letters), you'll get warnings from EvView when "Esc"
is pressed.

This is caused by the search being reset when exiting the preview. As we
don't have a document, or document model anymore, there's no search to
be done.

https://bugzilla.gnome.org/show_bug.cgi?id=759248